t016 green

+++ src/main/java/org/example/model/Article.java
@@ -8,14 +8,14 @@ import java.time.LocalDateTime;
 @Getter
 public class Article {
 
-    private int id;
+    private Long id;
     private String title;
     private String body;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
     private boolean blind;
 
-    private Article(int id, String title, String body, LocalDateTime createdDate, LocalDateTime modifiedDate, boolean blind) {
+    private Article(Long id, String title, String body, LocalDateTime createdDate, LocalDateTime modifiedDate, boolean blind) {
         this.id = id;
         this.title = title;
         this.body = body;
@@ -24,7 +24,7 @@ public class Article {
         this.blind = blind;
     }
 
-    public static Article of(int id, String title, String body, LocalDateTime createdDate, LocalDateTime modifiedDate, boolean blind){
+    public static Article of(Long id, String title, String body, LocalDateTime createdDate, LocalDateTime modifiedDate, boolean blind){
         return new Article(id, title, body, createdDate, modifiedDate, blind);
 
     }
